requirements.txt
aiogram==3.6.0
fastapi==0.115.0
uvicorn==0.30.6
SQLAlchemy==2.0.35
aiosqlite==0.20.0
pydantic-settings==2.4.0
python-dotenv==1.0.1
openai==1.52.0
stripe==10.9.0
httpx==0.27.2
itsdangerous==2.2.0
Pillow==10.4.0
